requirements.txt
streamlit
python-dotenv
spotipy